+++ CHANGES.md
@@ -3,6 +3,8 @@
 ## v1.1.1
 
 - Fixed the `convert` result from `base62x` to `buffer`, `hex`, `base64`, `utf8`.
+- Improved the sample `04-compare` with **binary-safe** tests.
+- Added supports for encoding `urlencode`.
 
 ## v1.1.0